+++ src/c4/yml/common.cpp
@@ -9,16 +9,34 @@ namespace c4 {
 namespace yml {
 
 #ifndef RYML_NO_DEFAULT_CALLBACKS
+
+void report_error_impl(const char* msg, size_t length, Location loc, FILE *f)
+{
+    if(!f)
+    {
+        f = stderr;
+    }
+    if(loc)
+    {
+        if(!loc.name.empty())
+        {
+            fprintf(f, "%.*s:", (int)loc.name.len, loc.name.str);
+        }
+        fprintf(f, "%zu:%zu:", loc.line, loc.col);
+        if(loc.offset)
+        {
+            fprintf(f, " (%zuB):", loc.offset);
+        }
+    }
+    fprintf(f, "ERROR: %.*s\n", (int)length, msg);
+    fflush(f);
+}
+
 namespace {
 
-void error_impl(const char* msg, size_t length, void * /*user_data*/)
+void error_impl(const char* msg, size_t length, Location loc, void * /*user_data*/)
 {
-    fprintf(stderr, "%.*s\n", (int)length, msg);
-    fflush(stderr);
-    if(c4::is_debugger_attached())
-    {
-        C4_DEBUG_BREAK();
-    }
+    report_error_impl(msg, length, loc, nullptr);
     ::abort();
 }
 
@@ -28,7 +46,7 @@ void* allocate_impl(size_t length, void * /*hint*/, void * /*user_data*/)
     if(mem == nullptr)
     {
         const char msg[] = "could not allocate memory";
-        error_impl(msg, sizeof(msg)-1, nullptr);
+        error_impl(msg, sizeof(msg)-1, {}, nullptr);
     }
     return mem;
 }
@@ -38,6 +56,7 @@ void free_impl(void *mem, size_t /*length*/, void * /*user_data*/)
     ::free(mem);
 }
 } // empty namespace
+
 
 Callbacks::Callbacks()
     :
@@ -111,9 +130,9 @@ void set_memory_resource(MemoryResource* r)
     s_memory_resource = r ? r : &s_default_memory_resource;
 }
 
-void error(const char *msg, size_t msg_len)
+void error(const char *msg, size_t msg_len, Location loc)
 {
-    s_default_callbacks.error(msg, msg_len);
+    s_default_callbacks.error(msg, msg_len, loc);
 }
 
 } // namespace yml