+++ src/c4/yml/common.hpp
@@ -2,7 +2,7 @@
 #define _C4_YML_COMMON_HPP_
 
 #include <cstddef>
-#include <c4/common.hpp>
+#include <c4/substr.hpp>
 
 #define RYML_INLINE inline
 
@@ -26,24 +26,36 @@
 
 #if RYML_USE_ASSERT
 #   define RYML_ASSERT(cond) RYML_CHECK(cond)
-#   define RYML_ASSERT_MSG(cond, msg) RYML_CHECK_MSG(cond, msg)
+#   define RYML_ASSERT_MSG(cond, /*msg, */...) RYML_CHECK_MSG(cond, __VA_ARGS__)
 #else
 #   define RYML_ASSERT(cond)
-#   define RYML_ASSERT_MSG(cond, msg)
+#   define RYML_ASSERT_MSG(cond, /*msg, */...)
 #endif
 
 
-#define RYML_CHECK_MSG(cond, msg)                           \
-    if(!(cond))                                             \
-    {                                                       \
-        ::c4::yml::error(msg ": expected true: " #cond);    \
-    }
-
-#define RYML_CHECK(cond)                            \
-    if(!(cond))                                     \
-    {                                               \
-        ::c4::yml::error("expected true: " #cond);  \
-    }
+#ifndef RYML_DBG
+#   define RYML_CHECK(cond) if(!(cond)) { ::c4::yml::error("ERROR: expected true: " #cond); }
+#   define RYML_CHECK_MSG(cond, msg) if(!(cond)) { ::c4::yml::error(msg ": ERROR: expected true: " #cond); }
+#else
+#   define RYML_CHECK(cond)                             \
+        if(!(cond))                                     \
+        {                                               \
+            if(c4::is_debugger_attached())              \
+            {                                           \
+                C4_DEBUG_BREAK();                       \
+            }                                           \
+            ::c4::yml::error(__FILE__ ":" C4_XQUOTE(__LINE__) ": ERROR: expected true: " #cond);  \
+        }
+#   define RYML_CHECK_MSG(cond, msg)                          \
+        if(!(cond))                                           \
+        {                                                     \
+            if(c4::is_debugger_attached())                    \
+            {                                                 \
+                C4_DEBUG_BREAK();                             \
+            }                                                 \
+            ::c4::yml::error(__FILE__ ":" C4_XQUOTE(__LINE__) ": ERROR: expected true: " #cond "\n" msg);  \
+        }
+#endif
 
 
 #pragma clang diagnostic pop
@@ -68,18 +80,68 @@ enum : size_t { npos = size_t(-1) };
 enum : size_t { NONE = size_t(-1) };
 
 //-----------------------------------------------------------------------------
+//-----------------------------------------------------------------------------
+//-----------------------------------------------------------------------------
+
+//! holds a position into a source buffer
+struct LineCol
+{
+    //!< number of bytes from the beginning of the source buffer
+    size_t offset;
+    //!< line
+    size_t line;
+    //!< column
+    size_t col;
+
+    LineCol() : offset(), line(), col() {}
+    //! construct from line and column
+    LineCol(size_t l, size_t c) : offset(0), line(l), col(c) {}
+    //! construct from offset, line and column
+    LineCol(size_t o, size_t l, size_t c) : offset(o), line(l), col(c) {}
+};
+
+
+//! a source file position
+struct Location : public LineCol
+{
+    csubstr name;
+
+    operator bool () const { return !name.empty() || line != 0 || offset != 0; }
+
+    Location() : LineCol(), name() {}
+    Location(                     size_t l, size_t c) : LineCol{   l, c}, name( ) {}
+    Location(csubstr n,           size_t l, size_t c) : LineCol{   l, c}, name(n) {}
+    Location(csubstr n, size_t b, size_t l, size_t c) : LineCol{b, l, c}, name(n) {}
+    Location(const char *n,           size_t l, size_t c) : LineCol{   l, c}, name(to_csubstr(n)) {}
+    Location(const char *n, size_t b, size_t l, size_t c) : LineCol{b, l, c}, name(to_csubstr(n)) {}
+};
+
+
+//-----------------------------------------------------------------------------
 
 /** the type of the function used to report errors. This function must
  * interrupt execution, either by raising an exception or calling
  * std::terminate()/std::abort(). */
-using pfn_error = void (*)(const char* msg, size_t msg_len, void *user_data);
+using pfn_error = void (*)(const char* msg, size_t msg_len, Location location, void *user_data);
 
-void error(const char *msg, size_t msg_len);
-
+/** trigger an error: call the current error callback. */
+void error(const char *msg, size_t msg_len, Location loc);
+/** @overload error */
+inline void error(const char *msg, size_t msg_len)
+{
+    error(msg, msg_len, Location{});
+}
+/** @overload error */
+template<size_t N>
+inline void error(const char (&msg)[N], Location loc)
+{
+    error(msg, N-1, loc);
+}
+/** @overload error */
 template<size_t N>
 inline void error(const char (&msg)[N])
 {
-    error(msg, N-1);
+    error(msg, N-1, Location{});
 }
 
 //-----------------------------------------------------------------------------
@@ -105,7 +167,7 @@ struct Callbacks
         void* mem = m_allocate(len, hint, m_user_data);
         if(mem == nullptr)
         {
-            this->error("out of memory");
+            this->error("out of memory", {});
         }
         return mem;
     }
@@ -115,15 +177,26 @@ struct Callbacks
         m_free(mem, len, m_user_data);
     }
 
+    void error(const char *msg, size_t msg_len, Location loc) const
+    {
+        m_error(msg, msg_len, loc, m_user_data);
+    }
+
     void error(const char *msg, size_t msg_len) const
     {
-        m_error(msg, msg_len, m_user_data);
+        m_error(msg, msg_len, {}, m_user_data);
+    }
+
+    template<size_t N>
+    inline void error(const char (&msg)[N], Location loc) const
+    {
+        error(msg, N-1, loc);
     }
 
     template<size_t N>
     inline void error(const char (&msg)[N]) const
     {
-        error(msg, N-1);
+        error(msg, N-1, {});
     }
 
 };